Restructure NotificationCenter and exclude test case
